Added a UDUnits::Convert() function that handles doubles for CAM data conversions, and directed the ROMS converter at this function instead of the float based implementation

+++ include/vapor/NetCDFCFCollection.h
@@ -275,8 +275,8 @@ public:
  virtual int Convert(
 	const string from,
 	const string to,
-	const float *src,
-	float *dst,
+	const double *src,
+	double *dst,
 	size_t n
  ) const;
